requirements.txt
PyGithub==1.55
PyInquirer==1.0.3